Telegram_Bot/pip-requirements.txt
python-telegram-bot==11.1.0
sentry_sdk==0.15.1
